Move Heatmap to appropriate package
