+++ src/org/labkey/test/pages/test/CoreComponentsTestPage.java
@@ -5,7 +5,7 @@ import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.WebTestHelper;
 import org.labkey.test.components.glassLibrary.components.ReactSelect;
 import org.labkey.test.components.glassLibrary.grids.EditableGrid;
-import org.labkey.test.components.glassLibrary.grids.HeatMap;
+import org.labkey.test.components.glassLibrary.heatmap.HeatMap;
 import org.labkey.test.components.glassLibrary.grids.QueryGrid;
 import org.labkey.test.components.html.Input;
 import org.labkey.test.components.ui.EntityInsertPanel;
