+++ src/org/labkey/test/tests/component/HeatMapTest.java
@@ -8,7 +8,7 @@ import org.junit.experimental.categories.Category;
 import org.labkey.test.BaseWebDriverTest;
 import org.labkey.test.TestTimeoutException;
 import org.labkey.test.categories.InDevelopment;
-import org.labkey.test.components.glassLibrary.grids.HeatMap;
+import org.labkey.test.components.glassLibrary.heatmap.HeatMap;
 import org.labkey.test.pages.test.CoreComponentsTestPage;
 import org.labkey.test.params.FieldDefinition;
 import org.labkey.test.params.experiment.SampleTypeDefinition;
